pyproject.toml
[project]
name = "resume-editor"
version = "0.60.19"
description = "Edits and refines resumes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "click>=8.2.1",
    "cryptography>=45.0.6",
    "fastapi>=0.119.0",
    "langchain-openai>=0.3.29",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-core<2.39",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "resume-writer",
    "sqlalchemy>=2.0.42",
    "starlette>=0.47.2",
    "typing-extensions>=4.14.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "aider-chat>=0.85.5",
    "bandit[baseline,toml]>=1.8.6",
    "beautifulsoup4>=4.13.4",
    "coverage>=7.10.3",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
]

[tool.pytest.ini_options]
pythonpath = [".", "resume_editor", "./tests", "../resume_writer"]
addopts = "-rx"
log_cli_level = "DEBUG"

[tool.uv.sources]
resume-writer = { git = "https://github.com/mpaguilar/resume_writer.git", rev = "44084d1" }

[tool.ruff]
exclude = ["temporal-samples-python"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "D201", "D202",
    "RET504", # ignore assignment right before return
    "PGH003", # ignore general mypy type: ignore
    "PTH123", # ignore Path.open
    "FA102", # ignore annotation warning, we require v3.10
    "ANN102", # ignore `cls` annotations
    "TD002", "TD003", "FIX002", # ignore TODO details
    "S101", # ignore asserts
    "ANN204", # ignore __init__ return type
    "UP032", "G001", # I like str.format
    "TRY003", # Ignore long exception messages
    "D104", "D100" # Ignore docstring for __init__.py, __main__.py
    ]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*" = ["ANN", "D", "I"]
